chore(API): implement refactor on controllers -created an answerCOntroller -restructure controllers conditional statement -remove comment within code [Started-#159923173]
